[KB] chore[#95]: Modal 디렉토리 기능별로 세부화
